+++ src/webapp/src/views/Dashboard.vue
@@ -2,12 +2,14 @@
 import { mapActions, mapState } from 'vuex'
 import Vue from 'vue'
 
+import EmbedShareButton from '@/components/generic/EmbedShareButton'
 import Report from '@/components/Report'
 import RouterViewLayout from '@/views/RouterViewLayout'
 
 export default {
   name: 'Dashboard',
   components: {
+    EmbedShareButton,
     Report,
     RouterViewLayout
   },
@@ -27,10 +29,6 @@ export default {
       'isLoadingActiveDashboard',
       'reports'
     ]),
-    dashboardEmail() {
-      // eslint-disable-next-line
-      return `mailto:?subject=Dashboard: ${this.activeDashboard.name}&body=${window.location}`
-    },
     displayedReports() {
       return this.isEditable || this.isUpdated
         ? this.editableDashboardReports
@@ -120,7 +118,7 @@ export default {
                 Cancel
               </button>
             </div>
-            <div v-else class="buttons is-right">
+            <div v-else class="buttons is-pulled-right">
               <a class="button" :href="dashboardEmail">Share</a>
               <button
                 v-if="!isEditable"
@@ -129,9 +127,10 @@ export default {
               >
                 Edit
               </button>
-              <router-link class="button" :to="{ name: 'dashboards' }"
-                >Back to Dashboards</router-link
-              >
+              <EmbedShareButton
+                :resource="activeDashboard"
+                resource-type="dashboard"
+              />
             </div>
           </div>
         </div>
@@ -147,15 +146,14 @@ export default {
           />
         </div>
 
-        <progress
-          v-else-if="isInitializing || isLoadingActiveDashboard"
-          class="progress is-small is-info"
-        ></progress>
-
         <div v-else class="columns">
           <div class="column">
             <div class="box">
-              <div class="content">
+              <progress
+                v-if="isInitializing || isLoadingActiveDashboard"
+                class="progress is-small is-info"
+              ></progress>
+              <div v-else class="content">
                 <p>No reports yet...</p>
               </div>
             </div>